+++ 4-period/java-vendingmachine-precourse-main/src/main/java/vendingmachine/domain/change/Change.java
@@ -1,4 +1,4 @@
-package vendingmachine.domain;
+package vendingmachine.domain.change;
 
 public class Change {
