+++ 4-period/java-vendingmachine-precourse-main/src/main/java/vendingmachine/domain/coin/Coin.java
@@ -1,10 +1,9 @@
-package vendingmachine;
+package vendingmachine.domain.coin;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import vendingmachine.domain.Coins;
 
 public enum Coin {
     COIN_500(500),
