+++ 4-period/java-vendingmachine-precourse-main/src/main/java/vendingmachine/domain/product/Product.java
@@ -1,4 +1,4 @@
-package vendingmachine.domain;
+package vendingmachine.domain.product;
 
 import vendingmachine.excpetion.InputException;
