+++ 4-period/java-vendingmachine-precourse-main/src/main/java/vendingmachine/domain/product/Products.java
@@ -1,4 +1,4 @@
-package vendingmachine.domain;
+package vendingmachine.domain.product;
 
 import java.util.HashMap;
 import java.util.List;
